+++ isf/testcase/add.rs
@@ -5,7 +5,7 @@ pub struct Add(u32);
 impl Default for Add {
     fn default() -> Self {
         let mut def = Self(0);
-        def.set_opcode(2u128.try_into().unwrap());
+        def.set_opcode(2u64.try_into().unwrap());
         def
     }
 }
@@ -52,15 +52,15 @@ impl Add {
         result.set_sign_extend(sign_extend.is_ok());
         let _ = winnow::ascii::multispace0.parse_next(input)?;
         let _ = "r".parse_next(input)?;
-        let dst: u128 = isf::parse::number_parser.parse_next(input)?;
+        let dst: u64 = isf::parse::number_parser.parse_next(input)?;
         result.set_dst(dst.try_into().unwrap());
         let _ = winnow::ascii::multispace0.parse_next(input)?;
         let _ = "r".parse_next(input)?;
-        let src1: u128 = isf::parse::number_parser.parse_next(input)?;
+        let src1: u64 = isf::parse::number_parser.parse_next(input)?;
         result.set_src1(src1.try_into().unwrap());
         let _ = winnow::ascii::multispace0.parse_next(input)?;
         let _ = "r".parse_next(input)?;
-        let src2: u128 = isf::parse::number_parser.parse_next(input)?;
+        let src2: u64 = isf::parse::number_parser.parse_next(input)?;
         result.set_src2(src2.try_into().unwrap());
         Ok(result)
     }
@@ -95,7 +95,7 @@ impl isf::MachineInstruction<u32> for Add {
     fn parse_machine(data: u32) -> Result<Self, isf::FieldMismatchError> {
         let perhaps = Self(data);
         let found = perhaps.get_opcode().try_into().unwrap();
-        let expected = 2u128;
+        let expected = 2u64;
         if found != expected {
             return Err(isf::FieldMismatchError {
                 field: "opcode".to_owned(),
