+++ isf/testcase/slice_add.rs
@@ -5,7 +5,7 @@ pub struct SliceAdd(u32);
 impl Default for SliceAdd {
     fn default() -> Self {
         let mut def = Self(0);
-        def.set_opcode(2u128.try_into().unwrap());
+        def.set_opcode(2u64.try_into().unwrap());
         def
     }
 }
@@ -49,10 +49,10 @@ impl SliceAdd {
         result.set_sign_extend(sign_extend.is_ok());
         let _ = winnow::ascii::multispace0.parse_next(input)?;
         let _ = "r".parse_next(input)?;
-        let dst: u128 = isf::parse::number_parser.parse_next(input)?;
+        let dst: u64 = isf::parse::number_parser.parse_next(input)?;
         result.set_dst(dst.try_into().unwrap());
         let _ = winnow::ascii::multispace0.parse_next(input)?;
-        let src: u128 = isf::parse::number_parser.parse_next(input)?;
+        let src: u64 = isf::parse::number_parser.parse_next(input)?;
         result.set_src(src.try_into().unwrap());
         Ok(result)
     }
@@ -83,7 +83,7 @@ impl isf::MachineInstruction<u32> for SliceAdd {
     fn parse_machine(data: u32) -> Result<Self, isf::FieldMismatchError> {
         let perhaps = Self(data);
         let found = perhaps.get_opcode().try_into().unwrap();
-        let expected = 2u128;
+        let expected = 2u64;
         if found != expected {
             return Err(isf::FieldMismatchError {
                 field: "opcode".to_owned(),
